MGMT-18913: Add support for external load balancer
This patch adds support for using an external load balancer. It will
only be available for OpenShift 4.16 or newer. To enable it the user
will have to use the new `load_balancer.type` field of the `cluster`
type, for example:

```json
{
  ...
  "load_balancer": {
    "type": "user-managed"
  }
}
```

To preserve backwards compatibility this will be optional, and the
default value will be `cluster-managed`.

When the value is `user-managed` the following will be added to the
generated `install-config.yaml` file:

```yaml
platform:
  baremetal:
    loadBalancer:
      type: UserManaged
```

Or for _vSphere_:

```
platform:
  vsphere:
    loadBalancer:
      type: UserManaged
```

The database will be automatically migrated so that for existing
clusters the value of the `load_balancer_type` column will automatically
be set to `cluster-managed`.

Note that this patch adds this support only for the `baremetal` and
`vsphere` platforms.

Related: https://issues.redhat.com/browse/MGMT-18913
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>